refactor(FeedbackModal): use general-purpose modal
